#24 [REFACTOR] explore, post 컴포넌트 폴더 구조화
