+++ src/components/explore/Cards/index.ts
@@ -1,0 +1,2 @@
+export { default as RecruitmentCard } from './RecruitmentCard';
+export { default as DesignerCard } from './DesignerCard';
