+++ src/components/explore/Filters/index.ts
@@ -1,0 +1,3 @@
+export { default as CategoryTabs } from './CategoryTabs';
+export { default as SubCategoryChips } from './SubCategoryChips';
+export { default as SortDropdown } from './SortDropdown';
