+++ src/components/explore/Header/SearchBar.tsx
