+++ src/components/explore/Header/index.ts
@@ -1,0 +1,2 @@
+export { default as ExploreHeader } from './ExploreHeader';
+export { default as SearchBar } from './SearchBar';
