+++ src/components/explore/index.ts
@@ -1,8 +1,4 @@
-export { default as ExploreHeader } from './ExploreHeader';
-export { default as CategoryTabs } from './CategoryTabs';
-export { default as SearchBar } from './SearchBar';
-export { default as SubCategoryChips } from './SubCategoryChips';
-export { default as SortDropdown } from './SortDropdown';
-export { default as RecruitmentCard } from './RecruitmentCard';
-export { default as DesignerCard } from './DesignerCard';
+export * from './Header';
+export * from './Filters';
+export * from './Cards';
