+++ src/components/post/Actions/index.ts
@@ -1,0 +1,1 @@
+export { default as PostActions } from './PostActions';
