+++ src/components/post/Content/DetailContent.tsx
