+++ src/components/post/Content/index.ts
@@ -1,0 +1,5 @@
+export { default as PostTabs } from './PostTabs';
+export { default as DetailContent } from './DetailContent';
+export { default as ServiceTags } from './ServiceTags';
+export { default as AvailableDates } from './AvailableDates';
+export { default as InfoSection } from './InfoSection';
