+++ src/components/post/Designer/index.ts
@@ -1,0 +1,1 @@
+export { default as DesignerInfo } from './DesignerInfo';
