+++ src/components/post/Gallery/index.ts
@@ -1,0 +1,1 @@
+export { default as ImageGallery } from './ImageGallery';
