+++ src/components/post/Header/index.ts
@@ -1,0 +1,1 @@
+export { default as PostHeader } from './PostHeader';
