+++ src/components/post/index.ts
@@ -1,10 +1,6 @@
-export { default as PostHeader } from './PostHeader';
-export { default as ImageGallery } from './ImageGallery';
-export { default as DesignerInfo } from './DesignerInfo';
-export { default as ServiceTags } from './ServiceTags';
-export { default as AvailableDates } from './AvailableDates';
-export { default as InfoSection } from './InfoSection';
-export { default as PostTabs } from './PostTabs';
-export { default as DetailContent } from './DetailContent';
-export { default as PostActions } from './PostActions';
+export * from './Header';
+export * from './Gallery';
+export * from './Designer';
+export * from './Content';
+export * from './Actions';
